CORE-6687: Simulator handles exceptions and timeouts in flows
Configuration has been added for a clock, flow timeout and poll interval for exception handling.
Defaults are set to default clock, 1 minute and 100 ms respectively.
FlowMessaging picks up error from Responder and gives it to Initiator, which wraps and rethrows.
A warning is included in error message that real Corda will not provide the cause.