CORE-6625: handle entityVersion consistently
In some places we did not retrieve entity version and so
could get optimistic concurrency exceptions from JPA.
This required changing the expected entity version level
assertions in the tests.

Refactor to combine CPI persist and update paths. The code for persist and update had a lot of similarities which
caused significant duplication and made it hard to read.

Make a number of entity fields val (i.e. readonly), since that is easier to reason about and maintain.

Have CpiMetadataEntity comply with the pattern that synthetic
primary keys are embedded, to make the code more uniform and
shorter.

Use the standard timestamp mechanism for dCpkDbChnageLogEntity

Note that insertTimestamp changes on update since that might
be surprising.